Drop comment_* apis.
Refactored dangling functions.
Added unit test.
